Validated and saving information in the correct folder
The previous version was also performing the failure test on the spot. I have decided to split the codes into one that saves the information (.h5 files and dictionary) and a different one that performs the failure test, which allows us to perform the test at any point, with any files.